Improve support for -Wundef on GCC.
Unfortunately, various parts of EASTL check preprocessor values
alternately as "#ifdef X" or "#if X". This means that if you build with
"-Wundef" on GCC, the "#if X" cases will cause warnings, but you cannot
fix them with "#define X 0" because that will cause wrong behavior when
the "#ifdef" method is used.

Such was the case for EA_COMPILER_CPP17_ENABLED, EA_COMPILER_CLANG, and
EA_COMPILER_GNUC.

This change causes these values to be checked in a more bullet-proof
fashion (though also more verbose).

Basically "#if X" has now become "#if (defined(X) && X)".